Reorganize plugin structure.
Let's take a more traditional route to this structure. The main "app"
file is now in the root of the plugin, the LEWP namespaced code is in
src, and vendor is at the root. Most composer packages look like this
and I'd rather conform to that standard than try to reinvent the wheel.

Additionally, having this code work as a plugin out of the box will make
contributing a heck of a lot easier.

My main reason for not doing this initially is that I loathe WP plugins
that ship with a bunch of dev stuff in it. My fix for this will be to
create a packaging build task that strips that junk out.