Bump the PHP minimum requirements to 5.4+.
I will be introducing a JWS and JWK package for signing requests in
later commits. In order to be able to use this package, the user must be
on PHP 5.4. This is now the minimum requirement. I'm not too concerned
about this at the moment, but if we want better support, we can rewrite
that library.