Fix build issues.
This commit removes phpcov and prefers the much simpler coverage clover
that is built directly into PHPUnit. All packages related to phpcov have
been removed.

The coverage report is now placed in "build/logs", which is apparently
the default and expected place for such a report. Let's not fight it and
just trust the default.

Our Travis configuration has also been updated to work better with
Coveralls. In other words, coverage reports are now being sent.

Finally, the phpunit.xml file was moved to the root as it is much easier
to maintain paths if it isn't buried in the "test" folder. It also now
has a whitelist to work with code coverage to only examine the right src
files.