Introduce Request and Directory classes.
The Request class is our base for all of the requests that are made when
interacting with the ACME RESTful API. Each type of request will extend
from it.

The Request class attempts to define the shared pieces of requests and
allow the extending class to implement custom pieces on top of it.

The Directory class is the first class that implements the Request
class. It simply makes a request to the ACME directory endpoint to
discover the RESTful resources for the particular ACME server.