Fix previous test case.
Intention is to test if the json_decode leads to a bad piece of data.
Previous version of the test never got to this place in code because the
"get" function returned an empty value. Let's make sure that the test
fails the json_decode, then make sure the code handles it properly.